requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
flask>=3.0.0
werkzeug>=3.0.0

